PolicyOps: move SingleLineBlock
